+++ package.json
@@ -10,6 +10,7 @@
         "@testing-library/user-event": "^12.1.10",
         "firebase": "^8.0.1",
         "framer-motion": "^2.9.4",
+        "js-base64": "^3.6.0",
         "node-sass": "4.14.1",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
