+++ src/App.scss
@@ -13,7 +13,7 @@
 @import "./Components/DashBoardContainer/SideBar/SideBar.scss";
 //AllRoute scss
 @import "./Components/DashBoardContainer/AllRoute/Profile/Profile.scss";
-@import "./Components/DashBoardContainer/AllRoute/MyOrder/MyOrder.scss";
+
 @import "./Components/DashBoardContainer/AllRoute/AddAdmin/AddAdmin.scss";
 @import "./Components/DashBoardContainer/AllRoute/AddFood/AddFood.scss";
 $mainColor: red;
