+++ src/Components/DashBoardContainer/AllRoute/AdminOrder/AdminOrder.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import DataTable from '../../DataTable/DataTable';
+
+const AdminOrder = () => {
+    //take all order for admin
+    return (
+        <div>
+            <DataTable controller={true}/>
+        </div>
+    );
+};
+
+export default AdminOrder;
