+++ src/Components/DashBoardContainer/OrderStepper/OrderStepperStyle.js
@@ -1,0 +1,93 @@
+import StepConnector from '@material-ui/core/StepConnector';
+import { makeStyles, withStyles } from '@material-ui/core/styles';
+
+const QontoConnector = withStyles({
+    alternativeLabel: {
+      top: 10,
+      left: 'calc(-50% + 16px)',
+      right: 'calc(50% + 16px)',
+    },
+    active: {
+      '& $line': {
+        borderColor: 'red',
+      },
+    },
+    completed: {
+      '& $line': {
+        borderColor: '#784af4',
+      },
+    },
+    line: {
+      borderColor: '#eaeaf0',
+      borderTopWidth: 3,
+      borderRadius: 1,
+    },
+  })(StepConnector);
+  
+ export const useQontoStepIconStyles = makeStyles({
+    root: {
+      color: 'red',
+      display: 'flex',
+      height: 22,
+      alignItems: 'center',
+    },
+    active: {
+      color: 'red',
+    },
+    circle: {
+      width: 8,
+      height: 8,
+      borderRadius: '50%',
+      backgroundColor: 'currentColor',
+    },
+    completed: {
+      color: 'red',
+      zIndex: 1,
+      fontSize: 18,
+    },
+  });
+
+  export const ColorlibConnectorStyle={
+    alternativeLabel: {
+      top: 22,
+    },
+    active: {
+      '& $line': {
+        background:'red',
+      },
+    },
+    completed: {
+      '& $line': {
+        background:'red',
+      },
+    },
+    line: {
+      height: 3,
+      border: 0,
+      backgroundColor: '#eaeaf0',
+      borderRadius: 1,
+    },
+  }
+
+  export const colorlibStepIconStyles={
+    root: {
+      backgroundColor: '#ccc',
+      zIndex: 1,
+      color: '#fff',
+      width: 50,
+      height: 50,
+      display: 'flex',
+      borderRadius: '50%',
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    active: {
+      backgroundImage:
+        'linear-gradient( 136deg, rgb(242,113,33) 0%, rgb(233,64,87) 50%, rgb(138,35,135) 100%)',
+      boxShadow: '0 4px 10px 0 rgba(0,0,0,.25)',
+    },
+    completed: {
+      backgroundImage:
+        'linear-gradient( 136deg, rgb(242,113,33) 0%, rgb(233,64,87) 50%, red 100%);',
+    },
+  }
