+++ src/Components/DashBoardContainer/SideBar/SideBarRoute/SideBarRoute.js
@@ -3,6 +3,7 @@ import {Switch,Route} from 'react-router-dom'
 import UserInfo from '../../../Authentication/UserInfo';
 import AddAdmin from '../../AllRoute/AddAdmin/AddAdmin';
 import AddFood from '../../AllRoute/AddFood/AddFood';
+import AdminOrder from '../../AllRoute/AdminOrder/AdminOrder';
 import MyOrder from '../../AllRoute/MyOrder/MyOrder';
 import OrderHistory from '../../AllRoute/OrderHistory/OrderHistory';
 import Profile from '../../AllRoute/Profile/Profile';
@@ -11,6 +12,9 @@ import Profile from '../../AllRoute/Profile/Profile';
 const SideBarRoute = () => {
     return (
         <Switch>
+            <Route exact path='/dashboard'>
+                 <MyOrder/>
+            </Route>
             <Route path='/dashboard/profile'>
                 <Profile/>
             </Route>
@@ -21,10 +25,10 @@ const SideBarRoute = () => {
                 <OrderHistory/>
             </Route>
             <Route path='/dashboard/logout'>
-                <UserInfo/>
+                <UserInfo showBtn={true}/>
             </Route>
             <Route path='/dashboard/admin/orders'>
-                 <OrderHistory/>
+                 <AdminOrder/>
             </Route>
             <Route path='/dashboard/add/admin'>
                <AddAdmin/>
